+++ src/main/java/ru/coffee/nostresso/controller/admin/AdminReviewController.java
@@ -20,28 +20,28 @@ public class AdminReviewController {
     }
 
     @GetMapping("/byShop")
-    public List<Review> getReviewsByShop(@RequestParam Integer shopId) {
+    public List<Review> getReviewsByShop(@RequestParam Long shopId) {
         return reviewService.findByShop(shopId);
     }
 
     @GetMapping("/byUser")
-    public List<Review> getReviewsByUser(@RequestParam Integer userId) {
+    public List<Review> getReviewsByUser(@RequestParam Long userId) {
         return reviewService.findByUser(userId);
     }
 
     @PostMapping
-    public Integer addReview(@RequestParam Integer shopId, @RequestBody Review review) {
+    public Long addReview(@RequestParam Long shopId, @RequestBody Review review) {
         return reviewService.addReview(shopId, review);
     }
 
     @PutMapping
-    public String updateReivew(@RequestParam Integer shopId, @RequestBody Review review) {
+    public String updateReivew(@RequestParam Long shopId, @RequestBody Review review) {
         reviewService.updateReview(shopId, review);
         return "review updated";
     }
 
     @DeleteMapping
-    public String deleteReview(@RequestParam Integer shopId, @RequestParam Integer reviewId) {
+    public String deleteReview(@RequestParam Long shopId, @RequestParam Long reviewId) {
         reviewService.deleteById(shopId, reviewId);
         return "review  " + reviewId + " deleted";
     }
