+++ src/main/java/ru/coffee/nostresso/model/mapper/OrderHistoryMapper.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 @Repository
 @Mapper
 public interface OrderHistoryMapper {
-    Optional<OrderHistory> findByUserId(Integer userId);
+    Optional<OrderHistory> findByUserId(Long userId);
 }
